Did some spring cleaning and moved lots of stuff
