+++ src/components/SignUpForm.jsx
@@ -4,7 +4,7 @@ import * as Yup from "yup";
 import ErrorMsg from "./ErrorMessage";
 import { GetResponseErrors } from "../utilities/CitysourcedResponseHelpers";
 import FormContainer from './FormContainer';
-import { SignUp } from './authService';
+import { SignUp } from '../services/authService';
 import { Link } from 'react-router-dom';
 import { IsFormComplete } from "../utilities/FormHelpers";
 import { IsPhoneNumberValid } from '@baltimorecounty/validation';
